# src/Microsoft.Identity.Web.Sidecar/OpenAPI/Microsoft.Identity.Web.Sidecar.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Microsoft.Identity.Web.Sidecar | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/Validate": {
      "get": {
        "tags": [
          "ValidateRequestEndpoints"
        ],
        "operationId": "ValidateAuthorizationHeader",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAuthorizationHeaderResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/AuthorizationHeader/{apiName}": {
      "get": {
        "tags": [
          "AuthorizationHeaderEndpoint"
        ],
        "summary": "Get an authorization header for a configured downstream API.",
        "description": "This endpoint will use the identity of the authenticated request to acquire an authorization header.Use dotted query parameters prefixed with 'optionsOverride.' to override call settings. Examples:\n  ?optionsOverride.Scopes=User.Read&optionsOverride.Scopes=Mail.Read\n  ?optionsOverride.RequestAppToken=true&optionsOverride.Scopes=https://graph.microsoft.com/.default\n  ?optionsOverride.AcquireTokenOptions.Tenant=common\nRepeat parameters like 'optionsOverride.Scopes' to add multiple scopes.",
        "operationId": "AuthorizationHeader",
        "parameters": [
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentIdentity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentUsername",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationHeaderResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/AuthorizationHeaderUnauthenticated/{apiName}": {
      "get": {
        "tags": [
          "AuthorizationHeaderEndpoint"
        ],
        "summary": "Get an authorization header for a configured downstream API using this configured client credentials.",
        "description": "This endpoint will use the configured client credentials to acquire an authorization header.Use dotted query parameters prefixed with 'optionsOverride.' to override call settings. Examples:\n  ?optionsOverride.Scopes=User.Read&optionsOverride.Scopes=Mail.Read\n  ?optionsOverride.RequestAppToken=true&optionsOverride.Scopes=https://graph.microsoft.com/.default\n  ?optionsOverride.AcquireTokenOptions.Tenant=common\nRepeat parameters like 'optionsOverride.Scopes' to add multiple scopes.",
        "operationId": "AuthorizationHeaderUnauthenticated",
        "parameters": [
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentIdentity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentUsername",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationHeaderResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/DownstreamApi/{apiName}": {
      "post": {
        "tags": [
          "DownstreamApiEndpoint"
        ],
        "summary": "Invoke a configured downstream API through the sidecar using the authenticated identity.",
        "description": "Override downstream call options using dotted query parameters prefixed with 'optionsOverride.'. Examples:\n  ?optionsOverride.Scopes=User.Read\n  ?optionsOverride.Scopes=User.Read&optionsOverride.Scopes=Mail.Read\n  ?optionsOverride.AcquireTokenOptions.Tenant=common\n  ?optionsOverride.RequestAppToken=true&optionsOverride.Scopes=https://graph.microsoft.com/.default",
        "operationId": "DownstreamApi",
        "parameters": [
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentIdentity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentUsername",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownstreamApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/DownstreamApiUnauthenticated/{apiName}": {
      "post": {
        "tags": [
          "DownstreamApiEndpoint"
        ],
        "summary": "Invoke a configured downstream API through the sidecar using the configured client credentials.",
        "description": "Override downstream call options using dotted query parameters prefixed with 'optionsOverride.'. Examples:\n  ?optionsOverride.Scopes=User.Read\n  ?optionsOverride.Scopes=User.Read&optionsOverride.Scopes=Mail.Read\n  ?optionsOverride.AcquireTokenOptions.Tenant=common\n  ?optionsOverride.RequestAppToken=true&optionsOverride.Scopes=https://graph.microsoft.com/.default",
        "operationId": "DownstreamApiUnauthenticated",
        "parameters": [
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentIdentity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AgentUsername",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownstreamApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthorizationHeaderResult": {
        "required": [
          "authorizationHeader"
        ],
        "type": "object",
        "properties": {
          "authorizationHeader": {
            "type": "string"
          }
        }
      },
      "DownstreamApiResult": {
        "required": [
          "statusCode",
          "headers",
          "content"
        ],
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "JsonNode": { },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ValidateAuthorizationHeaderResult": {
        "required": [
          "protocol",
          "token",
          "claims"
        ],
        "type": "object",
        "properties": {
          "protocol": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "claims": {
            "$ref": "#/components/schemas/JsonNode"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "ValidateRequestEndpoints"
    },
    {
      "name": "AuthorizationHeaderEndpoint"
    },
    {
      "name": "DownstreamApiEndpoint"
    }
  ]
}